Photo (re)naming feature.
In some photosets, pictures titles are duplicated, even if the pictures
are all different. Since this project was bound to the photo’s title
to create the file name, in photoset like this, after it downloaded and
stored the first photo, it would skip the next photos with the same name.

Now you can:
1. `append` the photo ID to its filename
2. `replace` the regular filename (photo title) with its ID
3. `increment` the filename (_photo.jpg, photo(1).jpg, etc._)
4. Don’t pass any __-n__ argument to perform the default (skip when same name)